# apps/api/openapi/moduflow-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Create a role in the active organization",
        "tags": [
          "Roles"
        ]
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List roles in the active organization",
        "tags": [
          "Roles"
        ]
      }
    },
    "/api/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Get a role by id in the active organization",
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Update a role in the active organization",
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          }
        },
        "summary": "Delete a role from the active organization",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/permissions/{id}": {
      "get": {
        "operationId": "PermissionsController_findOne",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Get a permission by id",
        "tags": [
          "Permissions"
        ]
      },
      "patch": {
        "operationId": "PermissionsController_update",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Update a permission",
        "tags": [
          "Permissions"
        ]
      },
      "delete": {
        "operationId": "PermissionsController_remove",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Delete a permission",
        "tags": [
          "Permissions"
        ]
      }
    },
...
    "/api/users/{userId}/roles/{roleId}": {
      "delete": {
        "operationId": "UserRolesController_removeRole",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the user.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRoleResponse"
                }
              }
            }
          }
        },
        "summary": "Remove a role from a user",
        "tags": [
          "User Roles"
        ]
      }
    },
    "/api/roles/{roleId}/permissions": {
      "post": {
        "operationId": "RolePermissionsController_assignPermission",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Assign a permission to a role",
        "tags": [
          "Role Permissions"
        ]
      },
      "get": {
        "operationId": "RolePermissionsController_findAll",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RolePermissionResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "List permissions assigned to a role",
        "tags": [
          "Role Permissions"
        ]
      }
    },
    "/api/roles/{roleId}/permissions/{permissionId}": {
      "delete": {
        "operationId": "RolePermissionsController_removePermission",
        "parameters": [
          {
            "name": "x-organization-id",
            "in": "header",
            "description": "Active organization identifier for platform-scoped routes.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the role.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "permissionId",
            "required": true,
            "in": "path",
            "description": "BigInt identifier for the permission.",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RolePermissionResponse"
                }
              }
            }
          }
        },
        "summary": "Remove a permission from a role",
        "tags": [
          "Role Permissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moduflow API",
    "description": "Moduflow API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RoleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Admin"
          },
          "description": {
            "type": "object",
            "example": "Can manage users and permissions.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Role name.",
            "example": "Admin",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Role description.",
            "example": "Can manage users and permissions."
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Role name.",
            "example": "Admin",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Role description.",
            "example": "Can manage users and permissions."
          }
        }
      },
...
      "PermissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "key": {
            "type": "string",
            "example": "users.manage"
          },
          "description": {
            "type": "object",
            "example": "Allows managing users.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "key",
          "description"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "Permission key.",
            "example": "users.manage",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Permission description.",
            "example": "Allows managing users."
          }
        }
      },
...
      "UserRoleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "name": {
            "type": "string",
            "example": "Admin"
          },
          "description": {
            "type": "object",
            "example": "Can manage users and permissions.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "AssignRolePermissionDto": {
        "type": "object",
        "properties": {
          "permissionId": {
            "type": "string",
            "description": "Permission id to assign to the role.",
            "example": "1"
          }
        },
        "required": [
          "permissionId"
        ]
      },
      "RolePermissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "key": {
            "type": "string",
            "example": "users.manage"
          },
          "description": {
            "type": "object",
            "example": "Allows managing users.",
            "nullable": true
          }
        },
        "required": [
          "id",
          "key",
          "description"
        ]
      },
...
    }
  }
}